LayoutTests/fullscreen/full-screen-document-background-color-expected.txt


EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK
RUN(enterFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(0, 128, 0)') OK
RUN(exitFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK
END OF TEST
